requirements-dev.txt
# Testing dependencies
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==20.1.0
freezegun==1.4.0

# Code quality
black==23.12.1
flake8==6.1.0
isort==5.13.2

# For mocking Docker and Celery
pytest-celery==0.0.0
